+++ changelog.md
@@ -1,5 +1,13 @@
 # sv-mosaic changelog
 
+## 1/30/2019
+* Renamed `Grid` component to `DataView`. `Grid` is still being exported but is considered deprecated. Update your implementations to import `DataView` as soon as possible. For consistency, several supporting components were re-named and should be updated from their deprecated names too:
+  * `GridFilterText` is now `DataViewFilterText`
+  * `GridFilterDropdown` is now `DataViewFilterDropdown`
+  * `GridFilterDropdownButtons` is now `DataViewFilterDropdownButtons`
+  * `GridFilterMultiselect` is now `DataViewFilterMultiselect`
+  * `GridPrimaryFilter` is now `DataViewPrimaryFilter`
+
 ## 11/12/2019
 * New Bundle - an updated bundle is now available.